requirements.txt
Flask==2.3.3
couchbase==4.1.5
confluent-kafka==1.9.2
watchdog==2.3.1
python-dotenv==1.0.0